requirements.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
django-filter==25.1